refactor(task-2): :building_construction: create app-routing
