+++ apps/my-app/src/app/app-routes.ts
@@ -1,0 +1,8 @@
+export const APP_ROUTES = {
+  HOME: 'home',
+};
+
+const BASE = "/";
+export const APP_ROUTES_GLOBAL = {
+  HOME: `${BASE}home`,
+};
