+++ apps/my-app/src/app/app.module.ts
@@ -1,41 +1,16 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-
-import {
-  PreloadAllModules,
-  RouteReuseStrategy,
-  RouterModule,
-  Routes
-} from '@angular/router';
-
-import {
-  IonicModule,
-  IonicRouteStrategy
-} from '@ionic/angular';
+import { IonicModule } from '@ionic/angular';
 
 import { AppComponent } from './app.component';
-
-const routes: Routes = [
-  {
-    path: 'home',
-    loadChildren: () => import('./home/home.module').then((m) => m.HomeComponentModule),
-  },
-  {
-    path: '/',
-    redirectTo: 'home',
-    pathMatch: 'full'
-  }
-];
+import { AppRoutingModule } from './app-routing.module';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [
     BrowserModule,
     IonicModule.forRoot(),
-    RouterModule.forRoot(routes, {preloadingStrategy: PreloadAllModules})
-  ],
-  providers: [
-    {provide: RouteReuseStrategy, useClass: IonicRouteStrategy}
+    AppRoutingModule
   ],
   bootstrap: [AppComponent],
 })
